+++ src/FakeItEasy/Creation/CastleDynamicProxy/CastleDynamicProxyGenerator.cs
@@ -168,23 +168,5 @@ namespace FakeItEasy.Creation.CastleDynamicProxy
                 return obj is InterceptEverythingHook;
             }
         }
-
-        private class ProxyInterceptor
-            : IInterceptor
-        {
-            private readonly IFakeCallProcessorProvider fakeCallProcessorProvider;
-
-            public ProxyInterceptor(IFakeCallProcessorProvider fakeCallProcessorProvider)
-            {
-                this.fakeCallProcessorProvider = fakeCallProcessorProvider;
-            }
-
-            public void Intercept(IInvocation invocation)
-            {
-                Guard.AgainstNull(invocation);
-                var call = new CastleInvocationCallAdapter(invocation);
-                this.fakeCallProcessorProvider.Fetch(invocation.Proxy).Process(call);
-            }
-        }
     }
 }
